+++ example/lib/widgets/app_drawer/app_drawer_tablet.dart
@@ -46,10 +46,9 @@ class AppDrawerTabletLandscape extends StatelessWidget {
           mainAxisAlignment: MainAxisAlignment.start,
           children: [
             SingleChildScrollView(
-              child: Column(
-                children: AppDrawer.getDrawerOptions(),
-              )
-            ),
+                child: Column(
+              children: AppDrawer.getDrawerOptions(),
+            )),
           ],
         ),
       ),