+++ lib/src/helpers/helpers.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:responsive_builder2/responsive_builder2.dart';
 
 // Author: https://github.com/fastogt/responsive_builder/tree/master