+++ lib/src/scroll/scroll_transform_view.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
 import 'package:responsive_builder2/src/scroll/scroll_transform_item.dart';