+++ test/helpers/helpers_test.dart
@@ -443,6 +443,7 @@ void main() {
           ),
         );
       }
+
       await testWithSize(
         tester: tester,
         size: const Size(4100, 1000),