+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - Add `ScreenTypeLayout.builder2` that includes `SizingInformation`.
  - Adds comparison operators to DeviceScreenType and RefinedSize, for convenience. (https://github.com/FilledStacks/responsive_builder/pull/55)
  - Code coverage reached 100%.
+ - Replace `import 'package:flutter/material.dart';` with `import 'package:flutter/widgets.dart';`.
 
 ## 0.8.2